# PERF: DataFrame floordiv