```
from asv_bench.benchmarks.arithmetic import *
self = MixedFrameWithSeriesAxis()
self.setup("floordiv")

%timeit self.time_frame_op_with_series_axis1("floordiv")
27.9 ms ± 987 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)  # <- master
23.7 ms ± 630 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)  # <- PR
```